Add files via upload
This is most of the backend for the database project.

Known errors:
Req. 9 - Still working on best way to fix. Logic error of Select.
Req. 10 - Not 100% on the requirements for this one. Unknown if we need to have automated way of determining if a paper has 2+ "Accepted" or not. Currently I have a field for "Accepted" Bool and believe that will satisfy for now.